requirements.txt
dnspython==2.2.1
grpcio==1.50.0
grpcio-tools==1.50.0
protobuf==4.21.9
pymongo==4.3.3
six==1.16.0
